langgraph/time-travel/pyproject.toml
[project]
name = "langgraph-time-travel"
version = "0.1.0"
description = "A LangGraph time travel example using Google AI"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.2.0",
    "langchain-google-genai>=2.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.9"

[tool.uv]
dev-dependencies = [] 